+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Added:
   * (#440) *TelegramBot#setChatStickerSet*, *TelegramBot#deleteChatStickerSet* (by @kamikazechaser)
 1. Support Bot API v3.5:
   * Support `provider_data` parameter in *TelegramBot#sendInvoice* (by @GochoMugo)
+1. Add options to *TelegramBot#stopPolling()* (by @GochoMugo)
 1. Add `metadata` argument in `message` event (and
    friends e.g. `text`, `audio`, etc.) (#409) (by @jlsjonas, @GochoMugo)
 1. Add forward-compatibility i.e. support future additional Telegram options (by @GochoMugo)